Add dynamic SEO metadata coverage
