+++ src/app/(main)/news/[id]/page.tsx
@@ -2,6 +2,12 @@ import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { getNewsOnlyItems, getNewsItemById } from "@/data/unified-news";
 import { NewsDetailContent } from "@/components/news/NewsDetailContent";
+import {
+  buildArticleJsonLd,
+  buildArticleMetadata,
+  buildNotFoundMetadata,
+  serializeJsonLd,
+} from "@/lib/seo";
 
 interface Props {
   params: Promise<{ id: string }>;
@@ -16,19 +22,15 @@ export async function generateStaticParams() {
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const { id } = await params;
   const item = getNewsItemById(id);
-  if (!item) return { title: "Not Found" };
+  if (!item) return buildNotFoundMetadata("News Not Found");
 
-  return {
+  return buildArticleMetadata({
     title: item.title,
     description: item.excerpt,
-    openGraph: {
-      title: item.title,
-      description: item.excerpt,
-      type: "article",
-      ...(item.image ? { images: [{ url: item.image }] } : {}),
-      ...(item.date ? { publishedTime: item.date } : {}),
-    },
-  };
+    path: `/news/${item.id}`,
+    image: item.image,
+    publishedTime: item.date,
+  });
 }
 
 export default async function NewsDetailPage({ params }: Props) {
@@ -39,6 +41,23 @@ export default async function NewsDetailPage({ params }: Props) {
   const relatedItems = getNewsOnlyItems()
     .filter((i) => i.id !== id)
     .slice(0, 3);
+  const jsonLd = buildArticleJsonLd({
+    title: item.title,
+    description: item.excerpt,
+    path: `/news/${item.id}`,
+    image: item.image,
+    publishedTime: item.date,
+  });
 
-  return <NewsDetailContent item={item} relatedItems={relatedItems} />;
+  return (
+    <>
+      <script
+        type="application/ld+json"
+        dangerouslySetInnerHTML={{
+          __html: serializeJsonLd(jsonLd),
+        }}
+      />
+      <NewsDetailContent item={item} relatedItems={relatedItems} />
+    </>
+  );
 }
